Cleanup started.
