+++ ethzasl_icp_mapper/launch/velodyne_puck/velodyne_puck__dynamic_mapper.launch
@@ -11,13 +11,11 @@ Parameters for ICRA 2014: "Long-term 3D map maintenance in dynamic environments"
 
 	<node name="mapper" type="dynamic_mapper" pkg="ethzasl_icp_mapper" output="screen" >
 		<remap from="cloud_in" to="/velodyne_points" />
-		<param name="subscribe_scan" value="false"/>
 		<param name="subscribe_cad" value="true"/>
 		<param name="icpConfig" value="$(find ethzasl_icp_mapper)/launch/velodyne_puck/icp_dynamic.yaml" />
 		<param name="inputFiltersConfig" value="$(find ethzasl_icp_mapper)/launch/velodyne_puck/input_filters.yaml" />
 		
 		<!--Put that back! -->
-		<param name="publishMapTf" value="false"/>
 		<param name="odom_frame" value="/world" /> #lidar
 
 		<param name="lidar_frame" value="lidar" />
